+++ src/config/constants.py
@@ -7,17 +7,17 @@ DEFAULT_STARTUP_TIMEOUT_SECONDS = 30.0
 DEFAULT_STARTUP_RETRY_ATTEMPTS = 3
 
 # Public namespace discovery endpoint.
-DEVELOPERS_NAMESPACE_LIST_URL = "https://developers.internal.nutanix.com/api/v1/namespaces"
+DEVELOPERS_NAMESPACE_LIST_URL = "https://developers.nutanix.com/api/v1/namespaces"
 
-# Internal YAML download endpoint pattern.
+# YAML download endpoint pattern.
 DEVELOPERS_YAML_DOWNLOAD_TEMPLATE = (
-    "https://developers.internal.nutanix.com/api/v1/namespaces/"
+    "https://developers.nutanix.com/api/v1/namespaces/"
     "{namespace}/versions/{version}/yaml"
 )
 
 # Namespace version listing endpoint used for latest-release artifact mode.
 DEVELOPERS_NAMESPACE_VERSIONS_TEMPLATE = (
-    "https://developers.internal.nutanix.com/api/v1/namespaces/{namespace}/versions"
+    "https://developers.nutanix.com/api/v1/namespaces/{namespace}/versions"
 )
 
 # Prism Central namespace version probe endpoint pattern.